updated registration route - add session setter function
