+++ src/app/api/auth/register/route.ts
@@ -1,17 +1,6 @@
 import { NextRequest, NextResponse } from 'next/server';
-import { setSession } from '../../../../lib/helpers/server_act_funcs/actions';
 import { psqlInsertUserCredentials, psqlIsEmailInUse } from '../../../../lib/sql/sqlQueries';
-import bcrypt from "bcrypt"
-
-async function hashPassword(password: string) {
-    try {
-        const hashedPassword = await bcrypt.hash(password, 10);
-        return hashedPassword;
-    } catch (err) {
-        console.error('Error hashing password:', err);
-        throw err;
-    }
-};
+import { hashPassword, setSessionCookie } from '../../../../lib/helpers/server_act_funcs/authorization';
 
 export async function POST(req: NextRequest) {
     const { username, email, password, role, birth_date, surname } = await req.json()
@@ -22,15 +11,15 @@ export async function POST(req: NextRequest) {
     // Signing up
     try {
         const isInUse = await psqlIsEmailInUse(email)
+        if (isInUse) return NextResponse.json({ message: "Email is already in use. Try different one" }, { status: 409 });
 
-        if (!isInUse) {
-            const hashedPassword = await hashPassword(password)
-            const user = await psqlInsertUserCredentials({ birth_date, email, password: hashedPassword, role, surname, username })
-            setSession(user)
-            return NextResponse.json({ message: "Successfully registered!" }, { status: 200 });
-        } else {
-            return NextResponse.json({ message: "Email is already in use. Try different one" }, { status: 409 });
-        }
+        const hashedPassword = await hashPassword(password)
+        const user = { birth_date, email, password: hashedPassword, role, surname, username }
+
+        await psqlInsertUserCredentials(user)
+        await setSessionCookie(user)
+
+        return NextResponse.json({ message: "Successfully registered!" }, { status: 200 });
     } catch (error) {
         return NextResponse.json({ message: "Unexpected error occured! Check console for details.", error: error }, { status: 409 });
     }
