+++ src/lib/helpers/server_act_funcs/actions.ts
@@ -1,8 +1,6 @@
 "use server";
 
 import { cookies } from "next/headers";
-import { AUTH_COOKIE_KEY } from "../../variables";
-import { ResponseCookie } from "next/dist/compiled/@edge-runtime/cookies";
 import { revalidateTag } from "next/cache";
 import {
   psqlAddUser,
@@ -26,17 +24,6 @@ export const storeThemeInCookies = (pref: string) => {
 // General
 export const readCookieForClient = async (searchCookie: string) => {
   return cookies().get(searchCookie)?.value;
-};
-export const setSession = async (token: string) => {
-  const options: Partial<ResponseCookie> = {
-    secure: true,
-    sameSite: "none",
-    httpOnly: true,
-    path: "/",
-  };
-
-  cookies().set(AUTH_COOKIE_KEY, token, options);
-  // cookies().set("user_id", user_id, options);
 };
 
 // Admin actions
